+++ fxn/beta/__init__.py
@@ -3,5 +3,5 @@
 #   Copyright © 2025 NatML Inc. All Rights Reserved.
 #
 
-from .metadata import CoreMLInferenceMetadata, GGUFInferenceMetadata, ONNXRuntimeInferenceMetadata
+from .metadata import CoreMLInferenceMetadata, GGUFInferenceMetadata, ONNXInferenceMetadata, ONNXRuntimeInferenceSessionMetadata
 from .remote import RemoteAcceleration